% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getconcor.R
\name{get.con.cor}
\alias{get.con.cor}
\title{Compute the conventional Pearson's correlation and concordance correlation
       of functional data}
\usage{
get.con.cor(X, Y)
}
\arguments{
\item{X:}{matrix with columns being time points and rows being subjects}

\item{Y:}{matrix with columns being time points and rows being subjects, same size as X}
}
\value{
rho: conventional Pearson's correlation coefficient

ccc: conventional concordance correlation coefficient
}
\description{
Compute the conventional Pearson's correlation coefficient and
             concordance correlation coefficeint of two data sets.
}
\examples{
x = matrix(rnorm(12), 3, 4)
y = matrix(rnorm(12), 3, 4)
get.con.cor(x,y)
}
