% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getcordata.R
\name{get.cor.data}
\alias{get.cor.data}
\title{Generate correlated data}
\usage{
get.cor.data(X, rho, weights, upper = 10000, round = T)
}
\arguments{
\item{X:}{matrix with columns being time points and rows being subjects}

\item{rho:}{functional Pearson's correlation coefficient}

\item{weights:}{a vector of weights whose length is number of time points}

\item{upper:}{the maximum value in the generated data set}

\item{round:}{If round is True, the generated data set will be rounded to the nearest integer.
By default, round equals True}
}
\value{
sim.data: correlated data set

rhoc: functional concordance correlation coefficient between X and sim.data
}
\description{
Given a data set X in matrix form, generate a new data set sim.data such that
             the functional Pearson's correlation between X and sim.data is rho. Also
             return the corresponding functional concordance correlation coefficient
}
\examples{
x = matrix(rnorm(12), 3, 4)
y = matrix(rnorm(12), 3, 4)
w = c(1,1,1,1)
get.cor.data(x, rho = 0.5, weights = w)
}
