% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfuncor.R
\name{get.fun.cor}
\alias{get.fun.cor}
\title{Compute the functional Pearson's correlation and concordance correlation
       of functional data}
\usage{
get.fun.cor(X, Y, W = NULL)
}
\arguments{
\item{X:}{matrix with columns being time points and rows being subjects}

\item{Y:}{matrix with columns being time points and rows being subjects, same size as X}

\item{W:}{weight matrix, same size as X and Y. By default, weight is the same for all data points}
}
\value{
rho: functional Pearson's correlation coefficient

ccc: functional concordance correlation coefficient

rho.se: standard error of functional Pearson's correlation coefficient

ccc.se: standard error of functional concordance correlation coefficient
}
\description{
Compute the functional Pearson's correlation coefficient and
             concordance correlation coefficeint of two data sets and the
             corresponding standard errors.
}
\examples{
x = matrix(rnorm(12), 3, 4)
y = matrix(rnorm(12), 3, 4)
w = matrix(1, 3, 4)
get.fun.cor(x,y,w)
}
